Golden & unified way to deploy falcon-sensor to the cloud
The script
 - (1) fetches customer credentials from AWS SSM parameter store
 - (2) downloads falcon sensor from CrowdStrike
 - (3) sets-up the sensor

I envision this script to be extended in future to be able to work with bundled
falcon-sensor and/or credentials outside of AWS SSM parameter store.

The script is compilation of existing prior work we had in this realm.
Thank You guys!

I think focusing on a single script that does the job well may turn out
beneficial in many ways. Single script means less documentation and single thing
to maintain.

This is POSIX script (tested with bash and dash). Minimal set-up of tools need
to be present on the system (curl, awk, grep). Tested on many platforms that are
supported by AWS SSM and Falcon, but not on all.

There is still work to be done on the script, but I think it is ready to be
merged.