feat (security hub): Add eventType filter to Stream data_feed in Security Hub (#240)
* feat (security hub): Add eventType filter to Stream data_feed in Security-Hub

This commit introduces a filter for `DetectionSummaryEvent` in the `data_feed` of the `Stream` class in the `Security-Hub` project. This filter is designed to reduce noise in the data stream by ensuring only relevant `DetectionSummaryEvent` types are included. The change improves the efficiency of our data stream processing by eliminating extraneous events.

* Update stream.py to remove redundant check of event type

* Updates installer to include new changes